4.21*: blogilistan laajennus, step9

+++ osa4/backend/controllers/blogs.js
@@ -27,6 +27,19 @@ blogsRouter.post("/", async (req, res) => {
 })
 
 blogsRouter.delete("/:id", async (req, res) => {
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return res.status(401).json({ error: "token invalid" })
+  }
+
+  const blog = await Blog.findById(req.params.id)
+  if (!blog) {
+    return res.status(204).end()
+  }
+  else if (decodedToken.id !== blog.user.toString()) {
+    return res.status(401).json({ error: "Unauthorized access" })
+  }
+
   await Blog.findByIdAndDelete(req.params.id)
   res.status(204).end()
 })
